@@ -5,8 +5,12 @@ import 'package:at_commons/at_commons.dart';
 import 'package:at_wavi_app/model/user.dart';
 import 'package:at_wavi_app/services/at_key_set_service.dart';
 import 'package:at_wavi_app/services/backend_service.dart';
+import 'package:at_wavi_app/services/nav_service.dart';
 import 'package:at_wavi_app/utils/at_enum.dart';
 import 'package:at_wavi_app/utils/at_key_constants.dart';
+import 'package:at_wavi_app/utils/constants.dart';
+import 'package:at_wavi_app/view_models/follow_service.dart';
+import 'package:provider/provider.dart';
 
 class AtKeyGetService {
   AtKeyGetService._();
@@ -18,7 +22,7 @@ class AtKeyGetService {
   init() {
     user = User(allPrivate: false, atsign: '');
     user.allPrivate = false;
-    user.atsign = BackendService().atClientInstance.currentAtSign!;
+    user.atsign = BackendService().atClientInstance.getCurrentAtSign()!;
   }
 
 // TODO: for testing only
@@ -34,7 +38,7 @@ class AtKeyGetService {
     bool _containsPrivateAccountKey = false;
     try {
       // _setUser(atsign: atsign);
-      atsign = atsign ?? BackendService().atClientInstance.currentAtSign;
+      atsign = atsign ?? BackendService().atClientInstance.getCurrentAtSign();
       var scanKeys = await BackendService().getAtKeys();
       user.allPrivate = true;
       for (var key in scanKeys) {
@@ -63,7 +67,7 @@ class AtKeyGetService {
   createPrivateAccountKey(
       String atsign, bool _containsPrivateAccountKey) async {
     try {
-      if (atsign == BackendService().atClientInstance.currentAtSign) {
+      if (atsign == BackendService().atClientInstance.getCurrentAtSign()) {
         if (!_containsPrivateAccountKey) {
           await AtKeySetService().update(
               BasicData(value: user.allPrivate.toString()),
@@ -89,6 +93,19 @@ class AtKeyGetService {
     }
 
     var successValue = await BackendService().atClientInstance.get(atKey);
+
+    if (atKey.key!.contains(MixedConstants.FOLLOWERS_KEY)) {
+      Provider.of<FollowService>(NavService.navKey.currentContext!,
+              listen: false)
+          .addFollowersData(successValue);
+    }
+
+    if (atKey.key!.contains(MixedConstants.FOLLOWING_KEY)) {
+      Provider.of<FollowService>(NavService.navKey.currentContext!,
+              listen: false)
+          .addFollowingData(successValue);
+    }
+
     if (successValue.value != null) {
       print('fetched value ${successValue.value} for key ${atKey.key}');
       isSetUserField = _setUserField(atKey.key, successValue.value, isCustom,
